pyproject.toml
[project]
name = "augment-atoms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ase>=3.25.0",
    "dacite>=1.9.2",
    "data2objects>=0.1.0",
    "load-atoms>=0.3.9",
    "vesin>=0.3.3",
]

[project.scripts]
augment-atoms = "augment_atoms:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/augment_atoms"]

[dependency-groups]
dev = [
    "bumpver>=2024.1130",
    "graph-pes>=0.0.36",
    "mace-torch>=0.3.12",
    "notebook>=7.4.0",
    "pytest>=8.3.5",
    "ruff>=0.11.6",
]

[tool.ruff]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM", "TID", "PIE"]
line-length = 79
respect-gitignore = true
target-version = "py39"
fix = true

[tool.bumpver]
current_version = "0.1.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
tag_scope = "global"
commit = true
tag = true
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"',
]
"src/augment_atoms/__init__.py" = [
    '__version__ = "{version}"',
]
